+++ tests/test_stats.py
@@ -2,26 +2,26 @@ import pytest
 import warnings
 
 import numpy as np
-import copia.richness as diversity
+import copia.estimators
 import copia.stats
 
 
 def test_bootstrap_seed():
     x = np.array([1, 1, 1, 2, 3, 5, 10, 25])
 
-    a = copia.stats.bootstrap(x, diversity.chao1,
+    a = copia.stats.bootstrap(x, copia.estimators.chao1,
               n_iter=1000, conf=0.95, n_jobs=4,
               seed=2021)
     
-    b = copia.stats.bootstrap(x, diversity.chao1,
+    b = copia.stats.bootstrap(x, copia.estimators.chao1,
               n_iter=1000, conf=0.95, n_jobs=4,
               seed=2021)
 
-    c = copia.stats.bootstrap(x, diversity.chao1,
+    c = copia.stats.bootstrap(x, copia.estimators.chao1,
               n_iter=1000, conf=0.95, n_jobs=4,
               seed=54656)
     
-    d = copia.stats.bootstrap(x, diversity.chao1,
+    d = copia.stats.bootstrap(x, copia.estimators.chao1,
               n_iter=1000, conf=0.95, n_jobs=4,
               seed=54656)
